new files: devide game factory interface and two concrete classes into multiple files
